+++ package.json
@@ -56,7 +56,8 @@
     "rdflib": "^1.0.6",
     "solid-auth-client": "^2.3.1",
     "solid-auth-tls": "^0.1.2",
-    "solid-namespace": "^0.3.0"
+    "solid-namespace": "^0.3.0",
+    "wrapper-webpack-plugin": "^2.1.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.6.2",
