+++ src/index.ts
@@ -59,6 +59,8 @@ import { infiniteMessageArea } from './chat/infinite'
 // @ts-ignore
 import pad from './pad'
 // @ts-ignore
+import participation from './participation'
+// @ts-ignore
 import preferences from './preferences'
 // @ts-ignore
 import store from './store'
@@ -91,6 +93,7 @@ if (typeof window !== 'undefined') {
     messageArea,
     infiniteMessageArea,
     pad,
+    participation,
     preferences,
     store,
     style,
